requirements.txt
django
django-bootstrap3
django-phonenumber-field>=0.2a3
django-template-debug
django-widget-tweaks
